# src/main/java/com/ariat/app/entity/WatchlistRequest.java
package com.ariat.app.entity;


import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class WatchlistRequest {
    private String stockName;
}