# src/main/java/com/ariat/app/entity/WatchlistResponse.java
package com.ariat.app.entity;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class WatchlistResponse {
    private String stockName;
    private StockDetails stockDetails;
}